requirements.txt
openvino
mediapipe
pyyaml